Add logging
Since program flow and minor errors are handled, in a big
part, with `Optional`, developers and users of the library
never really know what was going on in it, and had no way
to find probable causes of errors after they happened.

This commit introduces logging into the `jocument` library.
It uses the `log4j2` library (see more in the corresponding
issue #19).
Log statements were inserted into the program following my
verdict, since I do not really know of a good logging policy
(and did not find one).
Due to steering the program flow with optionals, many
execution branches which seem fatal are only logged with
`INFO` or `DEBUG` level since they actually occur pretty
often and are not really critical (see also in issue #19).

Signed-off-by: Anton Oellerer <a.oellerer@docu-tools.com>