Add custom placeholder infrastructure
To be able to add custom placeholders from outside of the
generator infrastructure allowing this should be created
(gh issue #26)
This commit introduces a first draft for it.
The way this works is by having a registry where custom
placeholder resolvers (of type `PlaceholderType.Custom`) can
be registered.
When the resolver is passed a placeholder to resolve, it
now consults this registry whether it has a resolver
for it registered, and if it does it returns an `Optional`
of the corresponding class.
The registry is passed to the resolver on first creation, and
subsequently gets passed to the recursive instantiations.

Signed-off-by: Anton Oellerer <a.oellerer@docu-tools.com>