Replace jlibvips with ImageStrategy Implementation
Instead of having jlibvips hardcoded and requirering all users of jocument to have this native dependency
we've implemented an ImageStrategy abstraction allowing users to implement their own strategy for image
processing in jocument.

To specify the ImageStrategy or any further divergences from the standard generation we've added GenerationOptions
which allow to specify customizations for Document generations upon starting the generation.

Closes #46