Fix mapping and parent resolving logic
Due to recursive parent resolving not existing in the
`FutureReflectionResolver`, bugs were uncovered when merging it into the
`ReflectionResolver`.
To fix this, the logic of when and where mapping and recursing into
parents is done has now be changed.
Basically, the flow now is
* can we resolve?
-> if no, map
-> can we resolve?
-> if no, recurse into the parent
quitting if we hit a yes along the way.

Additionally, the mapping is checked when trying to resolve
conditionals.