Return empty string if field is null
Up until now, if during reflection resolving a field was null,
the resolver continued search further up the object tree.
This should not happen, as a field being null also constitutes
information.
To fix this, from now on a ScalarPlaceholderData(null) is
returned, with the default formatting of null being the empty
string.