Fix Resolving IterablePlaceholders when no End-Marker present
When you have a placeholder that's being resolved to IterablePlaceholderData and
no End-Marker was present, the Word- and ExcelGenerators still tried to unroll
them. But the default behaviour should resolve a scalar value to the beans
toString method.

* Therefore, the WordGenerator#isLoopStart and ExcelGenerator#isLoopStart now
probe if a end loop marker is present and only then evaluate to true.
* The ReflectionResolver#toString method is returning the Beans toString value.