Migrate Excel creation to XSSF backend
To be able to operate with Excel templates
containing diagrams, the backend for Excel
generation is migrated to the XSSF one.
This required major refactorings in
`ExcelGenerator` and `ExcelWriter`.
The creation workflow is now similar to the
Word generation, as the template is modified,
loops inserted after the placeholder block, and
removing the block after finishing its
generation.