Fix Lost Cell Styles when Replacing Cells out of Loops
In ExcelGenerator#replacePlaceholder we've always replaced cells by new ones, which worked fine in loop bodies, but out of them it replaced the template cells on the spot and therefore got rid of the styles.

Now we differentiate between the two cases and replace the cell value instead of the cell itself.